requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.3
requests==2.32.3
bert-extractive-summarizer==0.10.1
torch==2.3.0 
torchvision==0.18.0 
torchaudio==2.3.0
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2